// libs/common-ui/src/lib/data/interfaces/dadata.interface.ts
export interface DadataSuggestion {
  value: 'г Санкт-Петербург';
  unrestricted_value: '190000, г Санкт-Петербург';
  data: {
    postal_code: '190000';
    country: 'Россия';
    country_iso_code: 'RU';
    federal_district: 'Северо-Западный';
    region_fias_id: 'c2deb16a-0330-4f05-821f-1d09c93331e6';
    region_kladr_id: '7800000000000';
    region_iso_code: 'RU-SPE';
    region_with_type: 'г Санкт-Петербург';
    region_type: 'г';
    region_type_full: 'город';
    region: 'Санкт-Петербург';
    area_fias_id: null;
    area_kladr_id: null;
    area_with_type: null;
    area_type: null;
    area_type_full: null;
    area: null;
    city_fias_id: 'c2deb16a-0330-4f05-821f-1d09c93331e6';
    city_kladr_id: '7800000000000';
    city_with_type: 'г Санкт-Петербург';
    city_type: 'г';
    city_type_full: 'город';
    city: 'Санкт-Петербург';
    city_area: null;
    city_district_fias_id: null;
    city_district_kladr_id: null;
    city_district_with_type: null;
    city_district_type: null;
    city_district_type_full: null;
    city_district: null;
    settlement_fias_id: null;
    settlement_kladr_id: null;
    settlement_with_type: null;
    settlement_type: null;
    settlement_type_full: null;
    settlement: null;
    street_fias_id: null;
    street_kladr_id: null;
    street_with_type: null;
    street_type: null;
    street_type_full: null;
    street: null;
    stead_fias_id: null;
    stead_cadnum: null;
    stead_type: null;
    stead_type_full: null;
    stead: null;
    house_fias_id: null;
    house_kladr_id: null;
    house_cadnum: null;
    house_flat_count: null;
    house_type: null;
    house_type_full: null;
    house: null;
    block_type: null;
    block_type_full: null;
    block: null;
    entrance: null;
    floor: null;
    flat_fias_id: null;
    flat_cadnum: null;
    flat_type: null;
    flat_type_full: null;
    flat: null;
    flat_area: null;
    square_meter_price: null;
    flat_price: null;
    room_fias_id: null;
    room_cadnum: null;
    room_type: null;
    room_type_full: null;
    room: null;
    postal_box: null;
    fias_id: 'c2deb16a-0330-4f05-821f-1d09c93331e6';
    fias_code: null;
    fias_level: '1';
    fias_actuality_state: '0';
    kladr_id: '7800000000000';
    geoname_id: '498817';
    capital_marker: '0';
    okato: '40000000000';
    oktmo: '40000000';
    tax_office: '7800';
    tax_office_legal: '7800';
    timezone: null;
    geo_lat: '59.939084';
    geo_lon: '30.315879';
    beltway_hit: null;
    beltway_distance: null;
    metro: null;
    divisions: null;
    qc_geo: '4';
    qc_complete: null;
    qc_house: null;
    history_values: ['г Ленинград'];
    unparsed_parts: null;
    source: null;
    qc: null;
  };
}